feat(sync): add SyncProtocolExecutor trait for unified protocol testing
Introduces a trait-based architecture for sync protocols that enables:
- Same protocol code to run in production and simulation
- Shared storage bridge (create_runtime_env) for both backends
- Standalone HashComparisonProtocol implementation

Changes:
- Add SyncProtocolExecutor trait in node-primitives
- Add create_runtime_env shared helper in storage_bridge.rs
- Extract HashComparisonProtocol to standalone module
- Clean up hash_comparison.rs (responder only, ~875 → ~285 lines)
- Fix wire protocol to use u64 for sequence_id (portability)
- Update simulation tests to use production protocol directly

This removes ~730 lines of duplicated code and ensures the simulation
tests exercise the exact same protocol logic as production.

+++ crates/node/primitives/src/sync/wire.rs
@@ -71,7 +71,7 @@ pub enum StreamMessage<'a> {
     /// Follow-up message in an ongoing sync operation.
     Message {
         /// Sequence number for ordering.
-        sequence_id: usize,
+        sequence_id: u64,
         /// The message payload.
         payload: MessagePayload<'a>,
         /// Nonce for the next message.
